+++ app/src/main/java/org/simple/clinic/help/HelpSync.kt
@@ -4,14 +4,15 @@ import io.reactivex.Completable
 import io.reactivex.Single
 import org.simple.clinic.sync.ModelSync
 import org.simple.clinic.sync.SyncConfig
+import org.simple.clinic.sync.SyncConfigType
+import org.simple.clinic.sync.SyncConfigType.Type.Daily
 import java.io.IOException
 import javax.inject.Inject
-import javax.inject.Named
 
 class HelpSync @Inject constructor(
     private val syncApi: HelpApi,
     private val syncRepository: HelpRepository,
-    @Named("sync_config_daily") private val config: SyncConfig
+    @SyncConfigType(Daily) private val config: SyncConfig
 ) : ModelSync {
 
   override val name: String = "Help"