+++ app/src/main/java/org/simple/clinic/sync/SyncConfigType.kt
@@ -6,6 +6,8 @@ import javax.inject.Qualifier
 annotation class SyncConfigType(val type: Type) {
 
   enum class Type {
-    Drugs
+    Daily,
+    Drugs,
+    Frequent
   }
 }